Revert "Prodduct view"
This reverts commit 60156b09aa79089efaf5637869c3f852380568fc.

+++ Common/ViewModels/Product/ProductAddViewModel.cs
@@ -12,9 +12,10 @@ namespace Common.ViewModels
         [Required(ErrorMessage = "Please input Name")]
         public string Name { get; set; }
         [Required(ErrorMessage = "Please input Price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [Required(ErrorMessage = "Please input Discount")]
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
+        //[Required(ErrorMessage ="Please select Image")]
         public string MainImage { get; set; }
         public ProductStatus  ProductStatus{ get; set; }
         public int CategoryId { get; set; }
